System: move gateway classes into module Domain folders
