+++ modules/User Admin/Domain/PersonGateway.php
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Gibbon\Domain\SchoolAdmin;
+namespace Gibbon\UserAdmin\Domain;
 
 use Gibbon\Domain\Gateway;
 use Gibbon\sqlConnection;
@@ -25,24 +25,27 @@ use Gibbon\Tables\DataFilters;
 use Gibbon\Tables\DataSet;
 
 /**
- * School Year Gateway
+ * Person Gateway
  *
  * Provides a data access layer for the gibbonSchoolYear table
  *
  * @version v16
  * @since   v16
  */
-class SchoolYearGateway extends Gateway
+class PersonGateway extends Gateway
 {
     public function selectAll(array $params)
     {
-        $totalRows = $this->pdo->executeQuery(array(), "SELECT COUNT(*) FROM gibbonSchoolYear")->fetchColumn(0);
+        $totalRows = $this->pdo->executeQuery(array(), "SELECT COUNT(*) FROM gibbonPerson")->fetchColumn(0);
         $params['totalRows'] = $totalRows;
 
         $filters = DataFilters::createFromArray($params);
 
         $data = array();
-        $sql = $this->applyDataFilters("SELECT * FROM gibbonSchoolYear", $filters);
+        // $sql = "SELECT gibbonPerson.*, gibbonRole.name as primaryRole FROM gibbonPerson LEFT JOIN gibbonRole ON (gibbonPerson.gibbonRoleIDPrimary=gibbonRole.gibbonRoleID)";
+        $sql = "SELECT gibbonPerson.gibbonPersonID, gibbonPerson.surname, gibbonPerson.preferredName, CONCAT(gibbonPerson.surname, ', ', gibbonPerson.preferredName) as fullName, gibbonPerson.username, gibbonPerson.image_240, gibbonPerson.status, gibbonRole.name as primaryRole FROM gibbonPerson LEFT JOIN gibbonRole ON (gibbonPerson.gibbonRoleIDPrimary=gibbonRole.gibbonRoleID)";
+        $sql = $this->applyDataFilters($sql, $filters);
+
         $result = $this->pdo->executeQuery($data, $sql);
 
         return DataSet::createFromResult($result)->setFilters($filters)->setTotalRows($totalRows);
